+++ src/milvus_project_helper/project.py
@@ -15,35 +15,11 @@ class ResourceNaming:
     user_password: str
 
 
-class PasswordStrengthError(Exception):
-    pass
-
-
 class ResourceExistsError(Exception):
     pass
 
 
-def check_password_strength(password: str):
-    """Check if the password is strong enough according to Milvus requirements."""
-    if len(password) < 8:
-        raise PasswordStrengthError("Password must be at least 8 characters long")
-    if not any(c.isupper() for c in password):
-        raise PasswordStrengthError(
-            "Password must contain at least one uppercase letter"
-        )
-    if not any(c.islower() for c in password):
-        raise PasswordStrengthError(
-            "Password must contain at least one lowercase letter"
-        )
-    if not any(c.isdigit() for c in password):
-        raise PasswordStrengthError("Password must contain at least one digit")
-    if not any(c in "!@#$%^&*()-+" for c in password):
-        raise PasswordStrengthError(
-            "Password must contain at least one special character"
-        )
-
-
-def project_create_resources(
+def create_resources(
     client: MilvusClient,
     resource_names: ResourceNaming,
     recreate_resources: bool = False,
@@ -134,10 +110,10 @@ def format_resource_status(
     """Format a resource status line with consistent symbols and indentation."""
     status_symbol = "✓" if exists else "×"
     status_color = "32" if exists else "31"  # 32=green, 31=red
-    return f"  \033[{status_color}m{status_symbol}\033[0m {resource_type}: {name}"
+    return f"  \033[{status_color}m{status_symbol}\033[0m {resource_type}: {name}"  # ]]
 
 
-def project_describe_resources(
+def describe_resources(
     client: MilvusClient, project_name: str, user_name: None | str = None
 ):
     """List the resources and collections in a project and check user privileges."""
@@ -189,17 +165,13 @@ def project_describe_resources(
     client.using_database("default")
 
 
-def project_drop_resources(
+def drop_resources(
     client: MilvusClient,
     project_name: str,
     database_name: None | str = None,
-    role_name: None | str = None,
-    user_name: None | str = None,
 ):
     """Drop all resources associated with a project."""
     database_name = database_name or f"db_{project_name}"
-    role_name = role_name or f"role_{project_name}"
-    user_name = user_name or f"user_{project_name}"
 
     database_exists = database_name in client.list_databases()
 
@@ -215,28 +187,22 @@ def project_drop_resources(
     # Switch to project database context
     client.using_database(database_name)
 
-    # Check and drop resources within database context
-    user_exists = user_name in client.list_users()
-    role_exists = role_name in client.list_roles()
-
-    logger.info(format_resource_status(user_name, user_exists, "user"))
-    logger.info(format_resource_status(role_name, role_exists, "role"))
-
     logger.info("\nDropping resources:")
 
-    if user_exists:
-        client.drop_user(user_name)
-        logger.info(f"  • Dropped user '{user_name}'")
+    for user_name in client.list_users():
+        if user_name != "root":
+            client.drop_user(user_name)
+            logger.info(f"  • Dropped user '{user_name}'")
 
-    if role_exists:
-        for priv in client.describe_role(role_name)["privileges"]:  # type: ignore
-            client.revoke_privilege(**priv)
-        client.drop_role(role_name)
-        logger.info(f"  • Dropped role '{role_name}'")
+    for role_name in client.list_roles():
+        if role_name not in ["admin", "public"]:
+            for priv in client.describe_role(role_name)["privileges"]:  # type: ignore
+                client.revoke_privilege(**priv)
+            client.drop_role(role_name)
+            logger.info(f"  • Dropped role '{role_name}'")
 
     # Switch back to default database to drop the project database
     client.using_database("default")
 
-    if database_exists:
-        client.drop_database(database_name)
-        logger.info(f"  • Dropped database '{database_name}'")
+    client.drop_database(database_name)
+    logger.info(f"  • Dropped database '{database_name}'")
