+++ src/milvus_project_helper/utils.py
@@ -1,0 +1,22 @@
+class PasswordStrengthError(Exception):
+    pass
+
+
+def check_password_strength(password: str):
+    """Check if the password is strong enough according to Milvus requirements."""
+    if len(password) < 8:
+        raise PasswordStrengthError("Password must be at least 8 characters long")
+    if not any(c.isupper() for c in password):
+        raise PasswordStrengthError(
+            "Password must contain at least one uppercase letter"
+        )
+    if not any(c.islower() for c in password):
+        raise PasswordStrengthError(
+            "Password must contain at least one lowercase letter"
+        )
+    if not any(c.isdigit() for c in password):
+        raise PasswordStrengthError("Password must contain at least one digit")
+    if not any(c in "!@#$%^&*()-+" for c in password):
+        raise PasswordStrengthError(
+            "Password must contain at least one special character"
+        )
